move long running tests to its own folder and category; check response and header for status/substatus
