+++ Microsoft.Azure.Cosmos/src/ChangeFeedProcessor/ChangeFeedEstimatorIterator.cs
@@ -290,8 +290,8 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed
                 //     And it is using the same lease container before the split occured.
                 //     And a ReadNextAsync returns a 410/1002 due to the partition no longer existing.
                 // Then the state of the change feed processor will have an estimatedLag equal to '1'.
-                if (response.CosmosException?.StatusCode == HttpStatusCode.Gone
-                    && response.CosmosException?.SubStatusCode == (int)SubStatusCodes.PartitionKeyRangeGone)
+
+                if (response.StatusCode == HttpStatusCode.Gone && response.Headers.SubStatusCode == SubStatusCodes.PartitionKeyRangeGone)
                 {
                     return (new ChangeFeedProcessorState(existingLease.CurrentLeaseToken, 1, existingLease.Owner), response);
                 }
